refactor(functions): add folder and doc migrators to common
