+++ apps/functions/applications-migration/nsip-document-migration/index.js
@@ -1,4 +1,4 @@
-import { migrationNsipDocuments } from './src/nsip-document-migration.js';
+import { migrationNsipDocuments } from '../common/migrators/nsip-document-migration.js';
 
 /**
  * @param {import('@azure/functions').Context} context
